+++ NEWS.md
@@ -1,3 +1,11 @@
+# SpectriPy 0.1
+
+## Changes in 0.1.0
+
+- Add `compareSpectriPy` method and related parameter objects to perform spectra
+  similarity calculations in python.
+
+
 # SpectriPy 0.0
 
 ## Changes in 0.0.2